+++ .github/workflows/push.yml
@@ -2,7 +2,7 @@ name: vikelabs-push
 
 on:
   push:
-    branches: [robert] # customize this to your branch if needed
+    branches: [team-lead-dashboard] # customize this to your branch if needed
 
 jobs:
   build:
