+++ components/ProjectEditor/Customizers/ImageCustomizer.tsx
@@ -13,8 +13,8 @@ import {
   Wrap,
 } from "@chakra-ui/react";
 import React, { useCallback, useEffect, useState } from "react";
-import { ImageInfo } from "../../types";
-import FileUploader from "../FileUploader";
+import { ImageInfo } from "../../../types";
+import FileUploader from "../../FileUploader";
 
 const ImageCustomizer = ({
   finalRef,
