--- components/svg/nextjs.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-function NextJsIcon() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="48" height="48" viewBox="0 0 48 48">
-      <path
-        fill="#212121"
-        d="M18.974 31.5a1.5 1.5 0 11-3 0v-14a1.499 1.499 0 012.724-.867l16.777 23.704C40.617 36.709 44 30.735 44 24c0-11-9-20-20-20S4 13 4 24s9 20 20 20c3.192 0 6.206-.777 8.89-2.122L18.974 22.216V31.5zm10-15a1.5 1.5 0 113 0v13.84l-3-4.227V16.5z"
-      ></path>
-    </svg>
-  );
-}
-
-export default NextJsIcon;
