--- components/svg/react.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-
-function ReactIcon() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="50"
-      height="50"
-      fillRule="evenodd"
-      clipRule="evenodd"
-      viewBox="0 0 50 50"
-    >
-      <path d="M34.555 3.984c-.78-.026-1.596.11-2.414.375-1.637.53-3.352 1.565-5.112 2.996a31.739 31.739 0 00-2.084 1.883 31.578 31.578 0 00-2.039-1.836c-1.758-1.43-3.468-2.464-5.101-2.994-1.633-.53-3.258-.546-4.592.225-1.334.77-2.132 2.184-2.49 3.863-.358 1.679-.319 3.678.04 5.916.126.778.304 1.593.507 2.424-.763.231-1.503.474-2.19.742-2.072.81-3.787 1.778-5.035 2.93-1.249 1.152-2.055 2.55-2.055 4.086 0 1.535.806 2.932 2.055 4.084C5.293 29.83 7.008 30.8 9.08 31.609c.734.287 1.53.545 2.35.79a32.188 32.188 0 00-.698 3.187c-.36 2.24-.397 4.24-.039 5.922.36 1.682 1.157 3.105 2.498 3.879 1.341.774 2.973.755 4.61.224 1.636-.53 3.35-1.562 5.11-2.994a31.643 31.643 0 002.066-1.863c.695.677 1.389 1.314 2.076 1.873 1.758 1.43 3.47 2.463 5.103 2.992 1.633.53 3.258.546 4.592-.224 1.334-.77 2.13-2.185 2.488-3.864.358-1.678.319-3.676-.04-5.914a32.164 32.164 0 00-.7-3.197c.848-.25 1.668-.515 2.424-.81 2.072-.81 3.787-1.78 5.035-2.932 1.249-1.152 2.055-2.549 2.055-4.084 0-1.536-.806-2.934-2.055-4.086-1.248-1.152-2.963-2.12-5.035-2.93a30.001 30.001 0 00-2.223-.752c.205-.836.386-1.657.512-2.44.36-2.24.398-4.24.039-5.923-.359-1.683-1.159-3.103-2.5-3.877-.67-.387-1.414-.575-2.193-.602zm-.092 2.036c.49.013.907.13 1.264.335.714.413 1.241 1.215 1.523 2.536.282 1.32.272 3.104-.06 5.173-.114.709-.28 1.46-.467 2.229a46.398 46.398 0 00-6.676-1.072 48.192 48.192 0 00-3.674-4.518 29.432 29.432 0 011.947-1.76c1.626-1.322 3.165-2.222 4.45-2.638.642-.209 1.204-.299 1.693-.285zm-18.977.005c.492-.013 1.055.077 1.7.286 1.288.417 2.83 1.317 4.457 2.64a29.362 29.362 0 011.906 1.723 48.3 48.3 0 00-3.701 4.556 46.219 46.219 0 00-6.647 1.08 28.875 28.875 0 01-.465-2.216c-.332-2.071-.34-3.856-.058-5.18.282-1.324.814-2.133 1.535-2.549.36-.208.781-.326 1.273-.34zm9.49 6.118c.815.886 1.63 1.856 2.438 2.9A57.87 57.87 0 0025 14.99c-.833 0-1.652.02-2.46.055a45.06 45.06 0 012.437-2.902zM25 17.01c1.36 0 2.685.056 3.975.15a56.84 56.84 0 012.58 4.057 57.157 57.157 0 011.98 3.752 57.171 57.171 0 01-2.008 3.802 57.142 57.142 0 01-2.011 3.215c-1.457.122-2.966.19-4.516.19a54.15 54.15 0 01-4.62-.2 57.175 57.175 0 01-1.978-3.165 57.159 57.159 0 01-2.002-3.795 57.037 57.037 0 012.012-3.817 56.736 56.736 0 012.568-4.035A54.3 54.3 0 0125 17.01zm6.549.4c1.648.206 3.196.491 4.637.834a45.018 45.018 0 01-1.592 4.375c-.414-.798-.843-1.6-1.307-2.402a60.068 60.068 0 00-1.738-2.807zm-13.205.016c-.58.89-1.15 1.799-1.7 2.752-.469.813-.907 1.625-1.326 2.433a45 45 0 01-1.58-4.35 42.914 42.914 0 014.606-.835zm19.82 1.35c.709.214 1.394.438 2.022.683 1.91.746 3.42 1.63 4.4 2.533.98.904 1.404 1.766 1.404 2.602 0 .836-.425 1.698-1.404 2.601-.98.904-2.49 1.787-4.4 2.534-.698.272-1.465.52-2.262.753a49.523 49.523 0 00-2.158-5.466 48.431 48.431 0 002.398-6.24zm-26.361.01a48.324 48.324 0 002.367 6.177A49.526 49.526 0 0012 30.46a27.8 27.8 0 01-2.186-.732c-1.91-.747-3.42-1.63-4.4-2.534-.98-.903-1.404-1.765-1.404-2.601 0-.836.425-1.698 1.404-2.602.98-.904 2.49-1.787 4.4-2.533a27.487 27.487 0 011.989-.674zM25 20c-2.75 0-5 2.25-5 5s2.25 5 5 5 5-2.25 5-5-2.25-5-5-5zm-9.658 7.365c.42.813.858 1.628 1.33 2.446.369.639.746 1.261 1.129 1.873a42.765 42.765 0 01-3.819-.702c.382-1.167.838-2.38 1.36-3.617zm19.277 0c.524 1.24.98 2.455 1.363 3.625a42.87 42.87 0 01-3.83.7c.388-.62.771-1.25 1.145-1.897.468-.811.904-1.621 1.322-2.428zm-21.213 5.559a45.9 45.9 0 005.852.972 49.317 49.317 0 004.29 5.391 29.372 29.372 0 01-1.927 1.744c-1.626 1.323-3.167 2.22-4.451 2.637-1.284.416-2.243.361-2.957-.05-.714-.413-1.24-1.215-1.522-2.536-.281-1.32-.273-3.103.059-5.172.15-.939.376-1.944.656-2.986zm23.155.008a29.92 29.92 0 01.66 3.002c.332 2.07.34 3.855.058 5.18-.282 1.324-.81 2.134-1.531 2.55-.72.416-1.688.47-2.977.053-1.288-.418-2.83-1.317-4.457-2.64a29.44 29.44 0 01-1.941-1.759 49.237 49.237 0 004.32-5.418 46 46 0 005.868-.968zm-14.694 1.17c1.027.056 2.067.093 3.133.093 1.03 0 2.037-.035 3.031-.088a46.038 46.038 0 01-3.084 3.74 46.198 46.198 0 01-3.08-3.745z"></path>
-    </svg>
-  );
-}
-
-export default ReactIcon;
