--- components/svg/typescript.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function TsIcon() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="50" height="50" viewBox="0 0 50 50">
-      <path d="M45 4H5a1 1 0 00-1 1v40a1 1 0 001 1h40a1 1 0 001-1V5a1 1 0 00-1-1zM29 26.445h-5V42h-4V26.445h-5V23h14v3.445zm1.121 14.667v-4.158s2.271 1.712 4.996 1.712c2.725 0 2.62-1.782 2.62-2.026 0-2.586-7.721-2.586-7.721-8.315 0-7.791 11.25-4.717 11.25-4.717l-.14 3.704s-1.887-1.258-4.018-1.258-2.9 1.013-2.9 2.096c0 2.795 7.791 2.516 7.791 8.141.001 8.664-11.878 4.821-11.878 4.821z"></path>
-    </svg>
-  );
-}
-
-export default TsIcon;
